Refactor service submenu pages into service/ folder
