+++ src/routes/router.tsx
@@ -6,10 +6,10 @@ import Dashboard from '../pages/Dashboard';
 import Booking from '../pages/dashboard/Booking';
 import Invoices from '../pages/dashboard/Invoices';
 import Notifications from '../pages/dashboard/Notifications';
-import Product from '../pages/dashboard/Products';
 import Profile from '../pages/dashboard/Profile';
-import ServicePackage from '../pages/dashboard/ServicePackage';
-import Services from '../pages/dashboard/Services';
+import Product from '../pages/dashboard/service/Products';
+import ServicePackage from '../pages/dashboard/service/ServicePackage';
+import Services from '../pages/dashboard/service/Services';
 import Settings from '../pages/dashboard/Settings';
 import Login from '../pages/Login';
 
@@ -33,11 +33,11 @@ export const router = createBrowserRouter([
         element: <Dashboard />,
       },
       {
-        path: 'services',
+        path: 'service/services',
         element: <Services />,
       },
       {
-        path: 'ServicePackage',
+        path: 'service/ServicePackage',
         element: <ServicePackage />,
       },
       {
@@ -53,7 +53,7 @@ export const router = createBrowserRouter([
         element: <Notifications />,
       },
       {
-        path: 'products',
+        path: 'service/products',
         element: <Product />,
       },
       {
